+++ lib/kender/commands/rspec.rb
@@ -8,7 +8,10 @@ module Kender
     end
 
     def command
-      if defined?(ParallelTests)
+      if defined?(Knapsack)
+        knapsack_env = "CI_NODE_TOTAL=#{ENV['CI_NODE_TOTAL']} CI_NODE_INDEX=#{ENV['CI_NODE_INDEX']}"
+        "#{knapsack_env} bundle exec rake knapsack:rspec"
+      elsif defined?(ParallelTests)
         'bundle exec rake parallel:spec'
       else
         'bundle exec rspec'